// packages/errors/src/types.ts
export enum BasedErrorCode {
  // Parse Errors
  incorrectFieldType = 1000,
  incorrectNodeType = 1001,
  exceedsMaximum = 1002,
  subceedsMinimum = 1003,
  fieldDoesNotExist = 1004,
  incorrectFormat = 1005,
  referenceIsIncorrectType = 1006,
  valueAndDefault = 1007,
  defaultNotSupported = 1008,
  multipleOperationsNotAllowed = 1009,
  requiredFieldNotDefined = 1010,
  languageNotSupported = 1011,
  invalidJSON = 1012,
  noLanguageFound = 1013,
  cannotDeleteNodeFromModify = 1014,
  nestedModifyObjectNotAllowed = 1015,
  infinityNotSupported = 1016,
  invalidSchemaFormat = 1017,
  invalidProperty = 1018,

  // client/server errors
  FunctionError = 50001,
  AuthorizeFunctionError = 50002,
  NoOservableCacheAvailable = 50003,
  ObservableFunctionError = 50004,
  ObserveCallbackError = 50005,
  FunctionNotFound = 40401,
  FunctionIsNotObservable = 40402,
  FunctionIsObservable = 40403,
  FunctionIsStream = 40404,
  CannotStreamToObservableFunction = 40405,
  FunctionIsWrongType = 40406, // was 40402 in server
  AuthorizeRejectedError = 40301,
  InvalidPayload = 40001,
  PayloadTooLarge = 40002,
  ChunkTooLarge = 40003,
  UnsupportedContentEncoding = 40004,
  NoBinaryProtocol = 40005,
  LengthRequired = 41101,
  MethodNotAllowed = 40501,
  RateLimit = 40029,
  MissingAuthStateProtocolHeader = 40030,
  IncorrectAccessKey = 40031,
  Block = 90001,

  // Mutation error
  PrefixAlreadyInUse = 2000,
  CannotChangeFieldInStrictMode = 2001,
  CannotRemoveFieldInStrictMode = 2002,
  CannotMutateWithExistingData = 2003,
  CannotDeleteRoot = 2004,
  CannotChangeDefaultField = 2005,
  CannotRemoveLastProperty = 2006,
}
// ...
export type BasedErrorPayload =
  | {
      observableId: number
      route: BasedRoute
    }
  | {
      requestId: number
      route: BasedRoute
    }
  | {
      channelId: number
      route: BasedRoute
    }
  | {
      streamRequestId: number
      route: BasedRoute
    }
  | { route: BasedRoute }